feat: add FilterOptions component
The FilterOptions component displays filter options for the available course items in the application. The component receives an array of objects containing information on the title and filter of each option. Clicking an option updates the active filter state and calls the Isotope component to update the display of course items according to the selected filter. The component uses React Bootstrap's Row, Col, Nav, and NavItem components.