build: Objetivos section from Sobre page
ObjetivosSection component with a carousel that displays the objectives of Menina de UX. The component imports the AllImages utility, React, and bootstrap components such as Col, Container, and Row. It also imports the Link component from react-router-dom and the CarouselItem component from the components/Carousel directory. The carouselItems constant is an array of objects that contain the image link, title, and descriptions of each carousel item. Finally, the component returns two sections: the first section is a hero image with the title 'Objetivos da Menina de UX' and the second section displays the carousel items with a pink background.